Move recursive resolution code to RecursiveResolver struct
