+++ src/dns/server.rs
@@ -18,91 +18,6 @@ impl Server {
         Server { context }
     }
 
-    fn lookup(&self, qname: &str, qtype: QueryType, server: (&str, u16)) -> Result<DnsPacket> {
-        let socket = UdpSocket::bind(("0.0.0.0", 3400)).unwrap();
-        let mut packet = DnsPacket::new();
-
-        packet.header.id = random::<u16>();
-        packet.header.questions = 1;
-        packet.header.recursion_desired = true;
-        packet
-            .questions
-            .push(DnsQuestion::new(String::from(qname), qtype));
-
-        let mut req_buffer = BytePacketBuffer::new();
-        packet.write(&mut req_buffer).unwrap();
-        socket.send_to(&req_buffer.buf[0..req_buffer.head()], server)?;
-
-        let mut res_buffer = BytePacketBuffer::new();
-        socket.recv_from(&mut res_buffer.buf).unwrap();
-
-        DnsPacket::from_buffer(&mut res_buffer)
-    }
-
-    fn recursive_lookup(&self, qname: &str, qtype: QueryType) -> Result<DnsPacket> {
-        // For now we're always starting with *a.root-servers.net*.
-        let mut ns = "198.41.0.4".to_string();
-
-        // Loop until we resolve the lookup
-        loop {
-            println!(
-                "\tAttempting lookup of {:?} {} with ns {}",
-                qtype, qname, ns
-            );
-            let ns_copy = ns.clone();
-            let server = (ns_copy.as_str(), 53);
-            let mut response = self.lookup(qname, qtype.clone(), server)?;
-
-            // If we have answers and no errors or the name server tells us no, done
-            if (!response.answers.is_empty() && response.header.rescode == ResponseCode::NOERROR)
-                || response.header.rescode == ResponseCode::NXDOMAIN
-            {
-                match qtype {
-                    QueryType::A => {
-                        let mut cname_responses: Vec<DnsRecord> = Vec::new();
-                        for rec in &response.answers {
-                            if let DnsRecord::CNAME { ref host, .. } = *rec {
-                                let cname_resp = self.recursive_lookup(&host, QueryType::A)?;
-                                println!("Resolved CNAME: {:?}", &host);
-                                response.header.rescode = cname_resp.header.rescode;
-
-                                for a_rec in cname_resp.answers {
-                                    cname_responses.push(a_rec);
-                                    response.header.answers += 1;
-                                }
-                            };
-                        }
-                        response.answers.extend(cname_responses);
-                    }
-                    _ => {}
-                }
-
-                return Ok(response);
-            }
-
-            // Otherwise, find the next name server
-            // First, check if we have the next NS's A record
-            if let Some(new_ns) = response.get_resolved_ns(qname) {
-                ns = new_ns;
-                continue;
-            }
-
-            // If not, resolve the IP of the NS
-            let new_ns_name = match response.get_unresolved_ns(qname) {
-                Some(name) => name,
-                None => return Ok(response),
-            };
-
-            // Now, we have to recursively resolve this NS's IP address
-            let recursive_response = self.recursive_lookup(&new_ns_name, QueryType::A)?;
-            if let Some(new_ns) = recursive_response.get_random_a() {
-                ns = new_ns;
-            } else {
-                return Ok(response);
-            }
-        }
-    }
-
     pub fn run(&self) -> ! {
         let socket = UdpSocket::bind(("0.0.0.0", self.context.dns_port)).unwrap();
         let resolver = self.context.get_resolver(self.context.clone());
